+++ RELEASE_NOTES.md
@@ -3,4 +3,5 @@
 
 ## Version 0.0.0
 ### Changes
+* Use BaseInformationBlock class
 * Initial version